Update urbandictionary utility to search for specific words:
* swag juice. noun.

'A substance which is not just obtained, but inherited. Usually imparts a walk with a slight limp, a smile that stops you in your tracks, and causes lucious barbaric man fur to grow from the chin. Characteristics include breaking curves in all classes, untouchable golf scores, and pool shark abilities. It drips off the person, leaving trails of swag juice in hallways and puddles on furniture.'